리덕스 툴킷 적용, 리엑트 쿼리 API 호출 적용
