+++ front/package.json
@@ -6,6 +6,7 @@
         "@emotion/react": "^11.10.5",
         "@emotion/styled": "^11.10.5",
         "@reduxjs/toolkit": "^1.9.1",
+        "@tanstack/react-query": "^4.22.0",
         "@testing-library/jest-dom": "^5.14.1",
         "@testing-library/react": "^13.0.0",
         "@testing-library/user-event": "^13.2.1",
