+++ front/src/App.tsx
@@ -1,28 +1,18 @@
-import React, { useState } from 'react';
-import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Login from 'auth/Login';
-import Signup from 'auth/Signup';
-import Landing from 'Landing';
-import Todo from 'Todo';
+import React from 'react';
+import Router from 'Router';
+import { Provider } from 'react-redux';
+import { store } from 'Globalstate/store';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+const queryClient = new QueryClient();
 
 function App() {
-    const token = localStorage.getItem('token');
-
     return (
-        <BrowserRouter>
-            <Routes>
-                {token && (
-                    <>
-                        <Route path="/" element={<Landing />}></Route>
-                        <Route path="/todo" element={<Todo />}></Route>
-                    </>
-                )}
-
-                <Route path="/" element={<Landing />}></Route>
-                <Route path="/login" element={<Login />}></Route>
-                <Route path="/signup" element={<Signup />}></Route>
-            </Routes>
-        </BrowserRouter>
+        <QueryClientProvider client={queryClient}>
+            <Provider store={store}>
+                <Router />
+            </Provider>
+        </QueryClientProvider>
     );
 }
